rename channel_response_filter.py to channel_response.py
