+++ mt_metadata/timeseries/filters/__init__.py
@@ -3,7 +3,7 @@ from .fir_filter import FIRFilter
 from .pole_zero_filter import PoleZeroFilter
 from .time_delay_filter import TimeDelayFilter
 from .frequency_response_table_filter import FrequencyResponseTableFilter
-from .channel_response_filter import ChannelResponseFilter
+from .channel_response import ChannelResponseFilter
 
 
 __all__ = [
