+++ mt_metadata/timeseries/filters/channel_response.py
